Migrate remaining plugins JUnit3 test files to Jupiter (JUnit5)
Converts the last 12 real (non-example) JUnit3 test files across
assetmaint, ecommerce, lucene, and scrum to Jupiter, using the
JunitJupiterTest/JupiterTestHelper infrastructure: drop the JUnit3
(String name) constructor, implement JupiterTestHelper instead of
extending OFBizTestCase, annotate methods with @Test, and move
sources from src/main to src/test. Each testdef's <junit-test-suite>
becomes <jupiter-test-suite>.

@Order is kept only on scrum's multi-method files, where
JupiterTestSuite's pinned MethodOrderer.OrderAnnotation is load-bearing:
SprintTests depends on testRemoveSprintMember running before
testAddSprintMember to avoid a pre-existing de-dup bug in
assignPartyToWorkEffort. assetmaint's FixedAssetMaintTests methods are
verified independent (each creates and queries only its own records),
so @Order was dropped there; ecommerce's OrderNotificationTests and
lucene's single-method LuceneTests never needed it.

All four suites verified passing under testIntegration:
assetmainttests (3), ecommercetests (51), lucenetests (17), and
scrumtests (479).
